refactor: move get user handle to messenger lib utils folder
