+++ src/components/messenger/list/utils/utils.tsx
@@ -1,6 +1,5 @@
 import { monthsSince, fromNow } from '../../../../lib/date';
 import { User } from '../../../../store/channels';
-import { Wallet } from '../../../../store/authentication/types';
 
 export function lastSeenText(user): string {
   if (user.isOnline) {
@@ -41,18 +40,4 @@ export function sortMembers(members: User[], adminIds: string[]) {
     // Finally sort alphabetically by firstName
     return a.firstName!.localeCompare(b.firstName);
   });
-}
-
-export function getUserHandle(primaryZID: string, wallets: Wallet[]) {
-  if (primaryZID) {
-    return primaryZID;
-  }
-
-  const publicAddress = wallets?.[0]?.publicAddress;
-
-  if (publicAddress) {
-    return `${publicAddress.substring(0, 6)}...${publicAddress.substring(publicAddress.length - 4)}`;
-  }
-
-  return '';
 }
